Move output code to separate file
